refactor : storage 폴더 생성 및 관련 파일 이동
